+++ src/app/app.module.ts
@@ -4,6 +4,7 @@ import { BrowserModule } from '@angular/platform-browser';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { CartModule } from './cart/cart.module';
+import { CoreModule } from './core/core.module';
 import { ProductsModule } from './products/products.module';
 
 @NgModule({
@@ -15,7 +16,7 @@ import { ProductsModule } from './products/products.module';
     AppRoutingModule,
 
     ProductsModule,
-    CartModule,
+    CartModule
 
   ],
   providers: [],