+++ src/app/core/models/config.model.ts
@@ -1,0 +1,8 @@
+export class ConfigModel {
+  constructor(
+    id: string,
+    login: string,
+    email: string,
+    phone: string
+  ) {}  
+}