+++ src/app/core/services/gen-id.generator.ts
@@ -1,0 +1,6 @@
+export function* genId() {
+    let i = 1;
+    while (true) {
+        yield i++;
+    }
+}