+++ src/app/core/services/local-storage.service.ts
@@ -1,0 +1,4 @@
+import { InjectionToken } from '@angular/core';
+
+export const localStorageToken = new InjectionToken<string>("localStorage");
+export const localStorageProvider = { provide: localStorageToken, useValue: window.localStorage };